requirements.txt
# Existing requirements
cryptography>=39.0.0
pillow>=9.0.0
zstandard>=0.19.0

# New SOTA requirements
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.14.0
opencv-python>=4.7.0
scikit-image>=0.21.0
numpy>=1.24.0
numba>=0.57.0
cupy-cuda11x>=11.0.0  # For GPU acceleration
intel-extension-for-pytorch>=2.0.0  # For CPU optimization
lpips>=0.1.4  # Perceptual loss
timm>=0.9.0  # PyTorch image models
transformers>=4.30.0  # Hugging Face transformers
triton>=2.0.0  # For Flash Attention
einops>=0.6.0  # For tensor operations
